fix: provide the right size availability info, better mapping of error codes
