+++ package.json
@@ -44,7 +44,7 @@
     "@redhat-cloud-services/frontend-components-notifications": "3.2.5",
     "@redhat-cloud-services/frontend-components-utilities": "3.2.16",
     "@rhoas/api-mock": "0.33.2",
-    "@rhoas/app-services-ui-components": "1.38.3",
+    "@rhoas/app-services-ui-components": "1.38.6",
     "@rhoas/app-services-ui-shared": "0.16.0",
     "@storybook/addon-actions": "6.5.5",
     "@storybook/addon-essentials": "6.5.5",
@@ -101,6 +101,7 @@
     "ts-loader": "9.2.8",
     "tsconfig-paths-webpack-plugin": "3.5.2",
     "tslib": "2.3.1",
+    "type-fest": "2.13.0",
     "typescript": "4.6.3",
     "url-loader": "4.1.1",
     "webpack": "5.71.0",
