--- src/app/modules/OpenshiftStreams/dialogs/CreateInstance/CreateInstanceWithSizes/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateInstance";
