+++ src/app/modules/OpenshiftStreams/dialogs/CreateInstance/index.ts
